==> 01-Introduction.ipynb <==
# Introduction

I used a dataset from Kaggle that contains every National Football League (NFL) play
from every regular season game for the 2008-2016 seasons. Each entry in the dataset 
contains information on each play from the game clock time and yards gained on the play 
to more advanced metrics like win probability and expected points. As there are often more 
than a hundred plays for just a single NFL game, the number of plays throughout these seasons
was very large (300,000+).

Since the biggest concern in sports is winning, I chose to analyze different statistics, both traditional and non-traditional, and see how they relate to a team's winning percentage.
