==> 02-ImportTidy.ipynb <==
# Import and Tidy

import pandas as pd
import numpy as np

## Load Dataset

nfl_plays = pd.read_csv('/data/JakeOliver28/maxhorowitz/nflplaybyplay2009to2016/NFL Play by Play 2009-2016 (v3).csv', \
                       dtype={'down':np.float64, 'DefTwoPoint':object, 'BlockingPlayer':str, 'ScoreDiff':float})

nfl_plays.head()

## Assert Data Types

# Verifying the data types are what they should be 
assert (nfl_plays['Date'].dtype == object)
assert (nfl_plays['GameID'].dtype == int)
assert (nfl_plays['Drive'].dtype == int)
assert (nfl_plays['qtr'].dtype == int)
assert (nfl_plays['down'].dtype == float)
assert (nfl_plays['time'].dtype == object)
assert (nfl_plays['TimeUnder'].dtype == int)
assert (nfl_plays['TimeSecs'].dtype == float)
assert (nfl_plays['PlayTimeDiff'].dtype == float)
assert (nfl_plays['SideofField'].dtype == object)
assert (nfl_plays['yrdln'].dtype == float)
assert (nfl_plays['yrdline100'].dtype == float)
assert (nfl_plays['ydstogo'].dtype == int)
assert (nfl_plays['ydsnet'].dtype == int)
assert (nfl_plays['FirstDown'].dtype == float)
assert (nfl_plays['posteam'].dtype == object)
assert (nfl_plays['DefensiveTeam'].dtype == object)
assert (nfl_plays['desc'].dtype == object)
assert (nfl_plays['PlayAttempted'].dtype == int)
assert (nfl_plays['Yards.Gained'].dtype == int)
assert (nfl_plays['Touchdown'].dtype == int)
assert (nfl_plays['ExPointResult'].dtype == object)
assert (nfl_plays['Safety'].dtype == int)
assert (nfl_plays['Passer'].dtype == object)
assert (nfl_plays['Passer_ID'].dtype == object)
assert (nfl_plays['PassAttempt'].dtype == int)
assert (nfl_plays['AirYards'].dtype == int)
assert (nfl_plays['YardsAfterCatch'].dtype == int)
assert (nfl_plays['InterceptionThrown'].dtype == int)
assert (nfl_plays['Rusher'].dtype == object)
assert (nfl_plays['Rusher_ID'].dtype == object)
assert (nfl_plays['RushAttempt'].dtype == int)
assert (nfl_plays['Receiver'].dtype == object)
assert (nfl_plays['Receiver_ID'].dtype == object)
assert (nfl_plays['Reception'].dtype == int)
assert (nfl_plays['Sack'].dtype == int)
assert (nfl_plays['Challenge.Replay'].dtype == int)
assert (nfl_plays['ChalReplayResult'].dtype == object)
assert (nfl_plays['PenalizedTeam'].dtype == object)
assert (nfl_plays['ScoreDiff'].dtype == float)
assert (nfl_plays['AbsScoreDiff'].dtype == float)
assert (nfl_plays['HomeTeam'].dtype == object)
assert (nfl_plays['AwayTeam'].dtype == object)
assert (nfl_plays['Field_Goal_Prob'].dtype == float)
assert (nfl_plays['Touchdown_Prob'].dtype == float)
assert (nfl_plays['ExPoint_Prob'].dtype == float)
assert (nfl_plays['TwoPoint_Prob'].dtype == float)
assert (nfl_plays['EPA'].dtype == float)
assert (nfl_plays['BlockingPlayer'].dtype == object)
assert (nfl_plays['Home_WP_pre'].dtype == float)
assert (nfl_plays['Away_WP_pre'].dtype == float)
assert (nfl_plays['Home_WP_post'].dtype == float)
assert (nfl_plays['Away_WP_post'].dtype == float)
assert (nfl_plays['WPA'].dtype == float)

## Ensure Consistent Naming

# Make sure old and new acronyms for teams are not both present
nfl_plays[['posteam', 'DefensiveTeam', 'HomeTeam', 'AwayTeam']] = \
    nfl_plays[['posteam', 'DefensiveTeam', 'HomeTeam', 'AwayTeam']].replace('LA', 'STL')
nfl_plays[['posteam', 'DefensiveTeam', 'HomeTeam', 'AwayTeam']] = \
    nfl_plays[['posteam', 'DefensiveTeam', 'HomeTeam', 'AwayTeam']].replace('JAC', 'JAX')

# Create columns detailing both the team and the season
nfl_plays['Year'] = nfl_plays['Date'].str.split('-').str.get(0)
nfl_plays = nfl_plays[nfl_plays['Year'] != '2017']
nfl_plays['posteamYear'] = nfl_plays['posteam'].astype(str) + nfl_plays['Year'].astype(str)
nfl_plays['DefensiveTeamYear'] = nfl_plays['DefensiveTeam'].astype(str) + nfl_plays['Year'].astype(str)
nfl_plays['HomeTeamYear'] = nfl_plays['HomeTeam'].astype(str) + nfl_plays['Year'].astype(str)
nfl_plays['AwayTeamYear'] = nfl_plays['AwayTeam'].astype(str) + nfl_plays['Year'].astype(str)


## Fill Null Values in Useful Columns

nfl_plays[['ScoreDiff', 'AbsScoreDiff']] = \
    nfl_plays[['ScoreDiff', 'AbsScoreDiff']].fillna(method='ffill')

## Assert Values

Checking that the values in the dataset fit within their expected bounds
assert (nfl_plays['down'].all() >= 1 and nfl_plays['down'].all() <= 4)
assert (nfl_plays['yrdline100'].all() >= 1 and nfl_plays['yrdline100'].all() <= 100)
assert (nfl_plays['TimeSecs'].all() >= 0 and nfl_plays['TimeSecs'].all() <= 3600)
assert (nfl_plays['PlayTimeDiff'].all() >= 0 and nfl_plays['PlayTimeDiff'].all() <= 90)
assert (nfl_plays['Drive'].all() >= 1 and nfl_plays['Drive'].all() <= 50)
assert (nfl_plays['qtr'].all() >= 1 and nfl_plays['qtr'].all() <= 4)
assert (nfl_plays['TimeUnder'].all() >= 0 and nfl_plays['TimeUnder'].all() <= 15)
assert (len(nfl_plays['SideofField'].all()) <= 3)
assert (nfl_plays['yrdln'].all() >= 1 and nfl_plays['yrdln'].all() <= 50)
assert (nfl_plays['ydstogo'].all() >= 0 and nfl_plays['ydstogo'].all() <= 99)
assert (np.absolute(nfl_plays['ydsnet'].all()) <= 99)
assert (len(nfl_plays[nfl_plays['posteam'].notnull()]['posteam'].all()) <= 3)
assert (np.absolute(nfl_plays['Yards.Gained'].all()) <= 99)
assert (np.absolute(nfl_plays['AirYards'].all()) <= 99)
assert (np.absolute(nfl_plays['YardsAfterCatch'].all()) <= 99)
assert (len(nfl_plays['HomeTeam'].all()) <= 3)
assert (len(nfl_plays['AwayTeam'].all()) <= 3)



nfl_plays.to_csv('/data/JakeOliver28/cleaned_nfl_plays.csv')
