==> 06-Conclusion.ipynb <==
# Conclusion

One thing I enjoyed about this dataset was that it was a very raw form of NFL data. Whereas fans are often presented with the stats of a given player or team over the course of a game or even season, this dataset contained the data that would be used to calculate those statistics. This provided me with a decent amount of freedom when deciding where to go with this project, being able to analyze many different phases of the NFL game. While having such basic data did make it difficult to both make sense of the data and tidy the data into a meaningful form, it was certainly rewarding.

Analyzing the NFL dataset, there were a decent amount of statistics that didn't provide much
clarity when trying to see what correlates with winning. However, multiple statistics were 
found to be useful when looking at the winning percentages of NFL teams each year. A team's 
offensive EPA per play, defensive EPA per play, total passing yards, and percent of runs that 
were on-pace seemed to be good predictors for this data. Using these variables as predictors, 
a linear regression model proved to be the most accurate in predicting winning percentage.

If I were to expand on this project, I could analyze what statistics are correlated with 
player success. My project ignored the distinction between different players having different
impacts and affecting win probabilities as such. To explore this aspect of the NFL could yield
interesting results.
Similarly, I could analyze certain teams' winning percentages throughout different seasons.
Grouping by both year and season, I treated each team in each year as a distinct team. If I 
were to take a more holistic approach for each franchise, I could potentially gain a better 
understanding of teams' successes during one season as it relates to previous seasons.
